Move widgets and add sdata booleans to interactive arguments
